- Update README.md - Add missing files (oops) - Add a new feature to customize hidden headers - Adjust namespace to EasyPost.Scotch
