+++ EasyPost.Scotch.Tests/RecordingTests.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Scotch.Tests
+namespace EasyPost.Scotch.Tests
 {
     [TestClass]
     public class RecordingTests : IDisposable
