+++ EasyPost.Scotch/EasyPost.Scotch.csproj
@@ -4,6 +4,7 @@
         <Nullable>enable</Nullable>
         <TargetFrameworks>net5.0;net6.0;netcoreapp3.1</TargetFrameworks>
         <LangVersion>latestMajor</LangVersion>
+        <RootNamespace>EasyPost.Scotch</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
