+++ EasyPost.Scotch/HttpClients.cs
@@ -1,15 +1,8 @@
 using System.Collections.Generic;
 using System.Net.Http;
 
-namespace Scotch
+namespace EasyPost.Scotch
 {
-    public enum ScotchMode
-    {
-        None = 0,
-        Recording = 1,
-        Replaying = 2
-    }
-
     public static class HttpClients
     {
         private static List<string>? DefaultCredentialHeadersToHide => new()
@@ -19,17 +12,17 @@ namespace Scotch
             }
         };
 
-        public static HttpClient NewHttpClient(string cassettePath, ScotchMode mode, bool hideCredentials = false)
+        public static HttpClient NewHttpClient(string cassettePath, ScotchMode mode, bool hideCredentials = false, List<string>? headersToHide = null)
         {
-            return NewHttpClientWithHandler(new HttpClientHandler(), cassettePath, mode, hideCredentials);
+            return NewHttpClientWithHandler(new HttpClientHandler(), cassettePath, mode, hideCredentials, headersToHide);
         }
 
-        public static HttpClient NewHttpClientWithHandler(HttpMessageHandler innerHandler, string cassettePath, ScotchMode mode, bool hideCredentials = false)
+        public static HttpClient NewHttpClientWithHandler(HttpMessageHandler innerHandler, string cassettePath, ScotchMode mode, bool hideCredentials = false, List<string>? headersToHide = null)
         {
             switch (mode)
             {
                 case ScotchMode.Recording:
-                    return new HttpClient(new RecordingHandler(innerHandler, cassettePath, hideCredentials ? DefaultCredentialHeadersToHide : null));
+                    return new HttpClient(new RecordingHandler(innerHandler, cassettePath, hideCredentials ? (headersToHide ?? DefaultCredentialHeadersToHide) : null));
                 case ScotchMode.Replaying:
                     return new HttpClient(new ReplayingHandler(innerHandler, cassettePath));
                 case ScotchMode.None:
