+++ EasyPost.Scotch/HttpInteraction.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace EasyPost.Scotch
+{
+    public class HttpInteraction
+    {
+        public DateTimeOffset RecordedAt { get; set; }
+        public Request Request { get; set; }
+        public Response Response { get; set; }
+    }
+}
