+++ EasyPost.Scotch/ScotchMode.cs
@@ -1,0 +1,8 @@
+namespace EasyPost.Scotch;
+
+public enum ScotchMode
+{
+    None = 0,
+    Recording = 1,
+    Replaying = 2
+}
