+++ EasyPost.Scotch/Status.cs
@@ -1,0 +1,9 @@
+using System.Net;
+
+namespace EasyPost.Scotch;
+
+public class Status
+{
+    public HttpStatusCode Code { get; set; }
+    public string? Message { get; set; }
+}
